# Support LZ4 compression in optimized parquet writer