--- a/plugin/trino-hive/src/test/java/io/trino/plugin/hive/benchmark/BenchmarkHiveFileFormat.java
+++ b/plugin/trino-hive/src/test/java/io/trino/plugin/hive/benchmark/BenchmarkHiveFileFormat.java
@@ -102,6 +102,7 @@ public class BenchmarkHiveFileFormat
 
     @Param({
             "NONE",
+            "LZ4",
             "SNAPPY",
             "GZIP"})
     private HiveCompressionCodec compression;