## Description
Support LZ4 compression in optimized parquet writer

```
read    LINEITEM    SNAPPY  TRINO_OPTIMIZED_PARQUET   858.0MB/s ±    24.0MB/s ( 2.80%) (N = 20, α = 99.9%)
read    LINEITEM    LZ4     TRINO_OPTIMIZED_PARQUET  1075.8MB/s ±    23.5MB/s ( 2.18%) (N = 20, α = 99.9%)
write   LINEITEM    SNAPPY  TRINO_OPTIMIZED_PARQUET   137.4MB/s ±  8419.6kB/s ( 5.98%) (N = 20, α = 99.9%)
write   LINEITEM    LZ4     TRINO_OPTIMIZED_PARQUET   136.4MB/s ±  7074.2kB/s ( 5.07%) (N = 20, α = 99.9%)
```

<!-- Provide details that would help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues
Fixes https://github.com/trinodb/trino/issues/9142

Note: LZ4 compression scheme has been deprecated by parquet spec in favor of LZ4_RAW https://github.com/apache/parquet-format/blob/master/Compression.md#lz4
However, LZ4 is the more widely supported encoding in other engines. E.g. https://issues.apache.org/jira/browse/SPARK-43273

<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

( ) This is not user-visible or docs only and no release notes are required.
( ) Release notes are required, please propose a release note for me.
(x) Release notes are required, with the following suggested text:

```markdown
# Hive
* Support LZ4 compression in optimized parquet writer. ({issue}`9142`)

# Hudi, Delta Lake, Iceberg
* Support LZ4 compression in parquet writer. ({issue}`9142`)
```