@raunaqmorarka 2023-06-05
> I don't think we should add support for a deprecated compression scheme in the writer. It's fine to add it to the _reader_ for interoperability purposes, though.
> 
> We should add LZ4_RAW, instead. When it becomes more widely adopted in other engines, users can pick that format instead of any one of the currently supported formats (zlib, snappy, etc).

Deprecation in parquet-format by itself is meaningless. The only thing that matters is what is widely adopted by the popular engines and libraries. See https://github.com/ClickHouse/ClickHouse/issues/49141 for an example of what happens if we implement what the spec recommends.
Adding LZ4_RAW option in the writer would create a way for users to shoot themselves in the foot by creating data that Spark cannot read https://issues.apache.org/jira/browse/SPARK-43273