@martint 2023-06-05
> See https://github.com/ClickHouse/ClickHouse/issues/49141 for an example of what happens if we implement what the spec recommends.

That's not the same. They changed the format unconditionally, without any option to choose an alternative. Of course that's going to cause problems until the format is widely supported everywhere. It's the reason why we don't choose certain default (e.g, we've delayed using zstd for parquet as the default compression format since other engines took a while to add support for it)

> Adding LZ4_RAW option in the writer would create a way for users to shoot themselves in the foot by creating data that Spark cannot read https://issues.apache.org/jira/browse/SPARK-43273

That's no different than supporting zstd when none of the other engines did. As long as it's not the default option, it reasonable to do so. There are users who might want to use LZ4 for performance reasons and don't use Spark at all. For those that do use Spark, they should test to make sure their choice of format is appropriate for interoperability.