@raunaqmorarka 2023-06-05
> That's no different than supporting zstd when none of the other engines did. As long as it's not the default option, it reasonable to do so. There are users who might want to use LZ4 for performance reasons and don't use Spark at all. For those that do use Spark, they should test to make sure their choice of format is appropriate for interoperability.

Not wanting to give users a way to generate incompatible data is the reason why we stripped out V2 data pages and V2 encodings support from the optimized writer (otherwise we would have retained it behind a flag while keeping V1 the default). This is also why we are still writing timestamps as INT96 and decimals as FIXED_WIDTH_BYTE_ARRAY in hive connector (to keep old versions of hive/spark happy) without an option to switch to newer more efficient encoding in INT32/INT64.
The old default parquet writer in hive connector already allows writing LZ4, it's just unimplemented in the optimized writer. Performance gain of LZ4 is the same as LZ4_RAW, so for a user who was using LZ4 upto this point, it's hard to justify switching to the less adopted format just because that is written in the spec.