@martint 2023-06-05
> We want to support LZ4 

What's the driver for this? Why not wait until other engines support reading it before we add support for writing? It's not like there aren't other alternatives (zstd, zlib, snappy) one could use.

> We could consider having a config flag later 

That seems suboptimal in light of my question above. Unless there's a compelling and urgent need to support LZ4, this is just trading a short term improvement for future complexity and tech debt (more config options, having to go through a deprecation process, etc)