@raunaqmorarka 2023-06-06
> > We want to support LZ4
> 
> What's the driver for this? Why not wait until other engines support reading it before we add support for writing? It's not like there aren't other alternatives (zstd, zlib, snappy) one could use.

There's couple of reasons: 
Writing LZ4 was supported by the previous parquet writer, the new one doesn't support it.
For someone wanting to use lightweight compression, LZ4 maybe the better choice over SNAPPY (significantly faster decoding with somewhat worse compression ratio).

> > We could consider having a config flag later
> 
> That seems suboptimal in light of my question above. Unless there's a compelling and urgent need to support LZ4, this is just trading a short term improvement for future complexity and tech debt (more config options, having to go through a deprecation process, etc)

We're not introducing any new configs here, writing LZ4 is already supported outside of optimized parquet writer.
This issue might not resolve quickly because LZ4 was deprecated by parquet two years ago and we still have current state with other engines.
Adding the option of LZ4_RAW doesn't look straightforward and it doesn't map to any existing Hadoop codec. Maybe this will get easier after hive/hadoop dependency removals ?

I'm fine with dropping this as it's not urgent. But it might become a sticking point for some existing user of hive connector with LZ4 compression.