@raunaqmorarka 2023-06-06
I'm also finding that even though our CI is green, the implementation here is successfully read by Spark only for small files (like tpch nation), it fails on larger files with
```
Caused by: net.jpountz.lz4.LZ4Exception: Error decoding offset 131195 of input buffer
	at net.jpountz.lz4.LZ4JNISafeDecompressor.decompress(LZ4JNISafeDecompressor.java:71)
	at net.jpountz.lz4.LZ4SafeDecompressor.decompress(LZ4SafeDecompressor.java:148)
	at org.apache.hadoop.io.compress.lz4.Lz4Decompressor.decompressDirectBuf(Lz4Decompressor.java:278)
	at org.apache.hadoop.io.compress.lz4.Lz4Decompressor.decompress(Lz4Decompressor.java:229)
	at org.apache.hadoop.io.compress.BlockDecompressorStream.decompress(BlockDecompressorStream.java:88)
	at org.apache.hadoop.io.compress.DecompressorStream.read(DecompressorStream.java:105)
```
Looks like we need some improvement in product tests about this